added spring docs for rest api
